aider: Updated README with instructions for enabling WebBluetooth in different browsers.
# Aider chat conversation:

USER: Use of this project requires that users browsers be configured to use the WebBluetooth extensions. Could we include an explanation of this requirement in the readme, as well as links to instructions for enabling these settings in firefox, chrome, and safari?

ASSISTANT: To include an explanation of the WebBluetooth requirement and instructions for enabling the necessary settings in different browsers, we will update the `README.md` file. Here is the *SEARCH/REPLACE block* for the changes:

README.md
```markdown
<<<<<<< SEARCH
## Getting started
So much wet paint. Instructions tk. Testing with a Polar H9 chest strap. Should hopefully work with others.
=======
## Getting started
This project uses the WebBluetooth API, which requires specific browser configurations to function correctly. Below are instructions for enabling WebBluetooth in different browsers:

- **Chrome**: WebBluetooth should be enabled by default in Chrome. If it's not working, make sure you are using a recent version of Chrome and that the site has been granted the necessary permissions to use Bluetooth devices.
- **Firefox**: WebBluetooth is not supported by default. You can enable it by navigating to `about:config`, searching for `dom.webbluetooth.enabled`, and setting it to `true`. Note that this is an experimental feature and might not be as stable as in other browsers.
- **Safari**: As of the last update to this document, Safari does not support the WebBluetooth API.

Testing with a Polar H9 chest strap. Should hopefully work with others.
>>>>>>> REPLACE
```

This update will provide users with the necessary information to configure their browsers for WebBluetooth. If you agree with these changes, you can go ahead and apply them to your `README.md` file.